Lots of changes Changed names of files
Added a user management class to hold data on what users are on the
system, what their authentication hash is, what groups they are in, and
what groups are available.

Added a file object class to hold data about individual files or
directories. This information is, the file name, owner, group, and
permissions.

Started work on the local path tree, which will contain data on all the
files on the local server.