Globally Unique Names
Decided to make it so all file names must be globally unique. This is to
help speed up the coding, so I can get the project done on time. This
can be changed in the future. Because I am doing this, the LocalPathList
will be used.

Also completed the server for receiving a new predecessor.