Node Failure
Finished detecting if a node has failed. The current node, checks to see
if its successor is still alive. If it is, than it grabs newer and new
files from its successor, and holds them as backups. If the successor is
not alive, if notifies its two hop successor that there has been a
failure. They then exchange files, while a notification goes out to the
rest of the network, for the failure. If a nodes search table is
modified, they send out their own message to gather information on the
network to reconstruct their search table.