Pushed user management
Altered the clients and servers, such that when a new node enters the
network, the node count is increased, and the user name associated with
the new node is pushed out. For now, everyone's password hash is 1. Not
secure, but just trying to get it finished on time. This can be easily
altered, as the backbone is already in place, to have different hashes.