Added ClntGetFile and ClntUploadFile
Removed some unnecessary members of Client class that weren't really
relevant.

Added some initialization to UserManagement that was missing and causing
some null pointer exceptions.

Changed DistFileSystemMain to show how tasks get added to the client
class.  Hopefully it doesn't break anything, but it makes the code a
little cleaner. I wasn't sure how to test it to make sure everything was
compatible, so I guess I'm leaving that part up to you.  If it gets too
hectic trying to fix it all feel free to revert it to how it works for
you or send me an email.