Removed unnecessary members of client side classes
Fleshed out user interface. Got pretty far, but some big things still
missing such as what to do with a file that has been newly downloaded?

Writing pops up a file dialog to pick a file. It should default to the
home directory but it doesn't usually.  May have to alt-tab to see the
dialog.  Feel free to change it to command line if you like.