File Downloading Works
Altered a few items in clntgetfile

Altered TaskManager to output what task it is working on. This will be
used later since right now, for some reason uploading and downloading
work when calling run directly, but now when adding them to the task
manager.

Altered the Usermangement, there was a problem with usernames getting
overwritten, that is now fixed.

Altered main, commented out the switch on string. This only works in
java7 which some linux package managers do not support yet.

Altered test, added some commands for testing

Altered check position, there was an odd error that kept cropping up,
that is now fixed

Altered get servgetfile a lot. It now sends file properly.