Fixed heartbeat and started group management
The heartbeat would not fire if a server saw a lot of traffic. This has
been changed from a socket timeout to a timer. It also caused an
infinite loop which has been addressed.

The start of group management has been put in. The client and server have
been created, but not yet tested.